requirements.txt
flask
pybit
python-dotenv
streamlit